% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rhelp0}
\alias{rhelp0}
\title{Info about R packages 0}
\usage{
rhelp0(question)
}
\arguments{
\item{question}{A question about R packages.}
}
\value{
The content of the response.
}
\description{
Info about R packages 0
}
\examples{
rhelp0("exp()")
}
