% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rhelp1}
\alias{rhelp1}
\title{Info about R packages}
\usage{
rhelp1(question)
}
\arguments{
\item{question}{A question about R packages.}
}
\value{
The content of the response.
}
\description{
Info about R packages
}
\examples{
rhelp1("exp()")
}
