% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rhelp2}
\alias{rhelp2}
\title{Info about R packages copy}
\usage{
rhelp2(question)
}
\arguments{
\item{question}{A question about R packages.}
}
\value{
The content of the response.
}
\description{
Info about R packages copy
}
\examples{
rhelp2("exp()")
}
