% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rhelp3}
\alias{rhelp3}
\title{Info about R packages 3}
\usage{
rhelp3(question)
}
\arguments{
\item{question}{A question about R packages.}
}
\value{
The content of the response.
}
\description{
Info about R packages 3
}
\examples{
rhelp3("exp()")
}
