% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rhelp4}
\alias{rhelp4}
\title{Info about R packages 4}
\usage{
rhelp4(question)
}
\arguments{
\item{question}{A question about R packages.}
}
\value{
The content of the response.
}
\description{
Info about R packages 4
}
\examples{
rhelp4("exp()")
}
